+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
 * Wish lists support integrating with DTR's item database URLs.
+* Added the ability to search for items that have notes. is:hasnotes
 
 # 5.51.0 (2019-10-20)